Remove scan_range from scan_builder
This is the first step to isolate scan_builder as a proper "query builder", that supports full queriable fields (id, timestamp, and secondary indexes) but also removes the support for range scans.
Instead, range queries have much simpler mechanisms that do not allow merging scans by design.

Since the API for range queries needed to be changed, the expiry of pending transfers was used as the main use case for the change, allowing custom logic to be inserted during the scan.