CLI: Add --development
### Context

- Production Linux binaries (`-Dconfig=production`) use `constants.direct_io_required == true` to require that Direct IO is available. (Excluding data files created on tmpfs.) If Direct IO is not available, they will panic.
- Non-production Linux binaries (`-Dconfig=development`) allow non-direct-IO, and just print a warning.

### Considerations

- Direct IO is critical to safety, to ensure the data is properly durable when the write completes.
- But we can't guarantee that a user's development environment has access to Direct IO. For example, some [file systems](https://github.com/tigerbeetle/tigerbeetle/issues/1835) inherently don't support it.
- We don't want to have multiple release binaries for Linux (i.e. one for production and one for development).
- When safety and user experience are in conflict, we prioritize safety.

### Solution

Add a new CLI flag to `tigerbeetle format` and `tigerbeetle start`. When `--development` is set, the absence of Direct IO is downgraded from a panic to a warning.
When `--development` is *not* set, print an informative message explaining all of this.

I considered several other alternate names, e.g. `--allow-buffered-io`. But while that is more precise, it is less clear to the end user, who is not as familiar with our internals.