cfo: add orchestrator script
cfo.zig takes care of running the fuzzers, but someone needs to run cfo!

Long term, this _probably_ should be integrated into some kind of
infrastructure as code setup, but, to start small, we'll be running cfo
in a persistent tmux session. Rather then relying on systemd or some
such, we'll write our own tiny "process manager" --- the only thing we
really need is:

* running in a loop
* ensuring that there are no orphaned processes

and both are easy to achieve with just sh. Note that it has to be sh,
rather than Zig, as this is bootstrap code that needs to be run with
what we have on a machine built-in (hence, POSIX sh and not bash)