tidy: blob check vs shallow clone
We want to avoid large files in the repository. Naively listing just the
files won't work: if the file is added in one commit and removed in a
subsequent one, it will be absent from the working tree, but will still
be perpetually present in the Git history.

For this reason, tidy is smart --- it asks git itself to list large git
objects.

But GitHub outsmarts tidy, as it does a shallow clone by default.

Fix this treachery within treachery within treachery with a proper clone

See https://stackoverflow.com/a/6802238/1936422 for why this specific
depth.