lsm: speed up tree fuzzer somewhat
Currently it timeouts on CFO. I am not sure this'll be enough to get it
to pass, but it seems a nice change regradless, as it speeds up the
fuzzing 2X for me

./zig/zig build -Drelease fuzz --
    --seed=3167000064919571166 --events-max=100000 lsm_tree

info(fuzz): done in 16.707s

VS

info(fuzz): done in 35.953s

The problematic path here is scans, where our data structure under test
is an LSM tree, and our model is a hash map. Hash maps can't do scans!

The proper fix here would be to use a BTree as a model, but Zig's stdlib
doesn't have a BTree.

So instead, we optimize the existing code, making use of the following
observation:

Our result set is limited, so most of the keys from the model do not
belong to the result set (even if they belong to the original scan range).

So, instead of binary searching every time, do a quick check for the
value being to the left/to the right of the result set, and fallback to
binary search only when the value is in bounds.