MessagePool: Lift `connections_max` to runtime
`connections_max` accounts for all possible replicas (6) and standbys (6).
But at runtime, we know how many replicas and standbys exist (potentially <6).

This means that we can reduce the number of messages allocated based on how many replicas/standbys there actually are.

In the case of a single-replica cluster, this save 60 messages (60MiB):

- `5` from `Replica.do_view_change_from_all_replicas`
- `11` from `Connection.recv_message`
- `11 * 4` from `Connect.send_queue`

Weirdly, this causes the virtual memory to _increase_ (presumably due to the changed access pattern to the arena allocator).