lsm: correct wrong comment
As far as I can tell, the fact that we don't store key min is
irrelevant, as, in the original search, we are comparing target key_min
with table's key_max anyway.

What actually matters is that the underlying binary search uses the
upper bound regardless of direction.

That is, when we looking ASCENDING for 10 and our level contains 5, the
result of binary search is

| 5 |  |
      ^ insertion position for 10, next item is null

if we are looking  DESCENDING for 10 and our level contains 50, the
result of the search is

| 50 |  |
  ^ insertion position for 10, next item is 50

I am torn whether we should just pass `direction` to the `.search`
method, but I think it won't be an improvement, all things considered.