lsm: drop support for perf from benchmarking
While I really love the API this provides, where we can observe _parts_
of the process from within, I don't think it's the right time to
maintain this code:

* It is finicky and platform dependent --- it fails to compile if I
  includeLibC for some reason, it fails at runtime on Alpine, it requires
  opt-out of benchmarking on other OSes.
* At the moment, it doesn't provide _that_ much value compared to using
  perf extenraly on the entire process

But perhaps the biggest thing I am worrying about here is how this will
evolve with Zig --- I would be surprised if the exact same interface
survives until 1.0!

So, lets remove this. If we add this back, it would be important to make
sure that the code doesn't have built-time depenrency on Linux. That is,
running it on Windows should just print _less_ information, instead of
outright failing with a compilation error.