lsm_fuzz: speed it up
The core reason why it is slow is that it is using a HashMap as a model,
but, since then introduction of scans, we really need a data structure
supporting range queries.

While Zig's stdlib doesn't have such a datastructure, we do! We can use
SortedSegmentedArray.

The following seed now finishes in under 5 minutes on my machine:

Seed: ./zig/zig build -Drelease fuzz -- --seed=17470417632183688204 lsm_tree

I have a suspicion that this change might still not be enough, and that
we should change our event generation to just make scans more _costly_
(one scan is equivalent to scan_result_max puts), but lets postpone that
until CFO finds a failing seed.