cfo: prepare to fetch PRs to fuzz from GitHub
The main change here is making the list of fuzzing tasks a runtime-known
weighted slice. It is still populated from the main branch.

This also adds a 'branch' field to the JSON record of  seed. Branch is a
bit of a misnomer, it is actually
`branch_or_pull_request_url_on_the_upstream_forge`, but branch is close
enough.