Multiversion: make release_client_min a parameter, set by release.zig
There was a bit of back and forth thinking as to where it should be set.
Figured release.zig was the most natural place, although it does feel a bit
strange to have a hardcoded constant there, we need it _somewhere_ and tying
it in to the code directly felt worse off.