cfo: build all fuzzers first
For some reason, all VOPR runs timed out on CFO:

https://github.com/tigerbeetle/devhubdb/blob/9ee578f1f39530573dafcf5dfa1c856f312ef1bc/fuzzing/data.json#L530-L617

(note `"ok": false` and `"seed_timestamp_end": 1715094977`) for each one
of the seeds.

I wasn't able to reproduce this locally or on the CFO machine, but my
hypothesis here is that the build system deadlocks with a lot of
concurrency.

Restore the behavior of building all code _once_ before fuzzing, even if
we don't need some of the fuzzers.

Note that I've unintentionally didn't remove building of minor fuzzers,
so those _dont_ suffer concurrent build issues, and perhaps are not
timing out for this reason.