cfo: actually make file non-blocking
SETFD manipulates flags of the file _descriptor_ (a single file can have
many descriptors pointing to it).

SETFL manipulates flags of the file itself (they are shared between all
descriptors pointing at the file).

NonBlocking is sadly a property of file, and not of a file descriptor,
though this isn't a problem for our use-case.

What is a problem is that I wrongly used SETFD instead of SETFL, which
made this into a no-op (it worked in practice because the default pipe
buffer was enough to accomodate one byte a second).