cfo: vopr no longer hangs
Currently vopr hangs in the CFO. This due to the classical bug of
systems programming: VOPR logs to the stderr, but no one reads the
stderr, so it fills up and then the process deadlocks.

Good news is that we have a fix for these classes of programs: shell.zig
intends to wrap raw child process spawning API in a set of well-defined
high-level operations, which are protected from these kinds of footguns.

The bad news is that the actual API there is buggy, and defaults to
piping by default. Fix this and make each caller that wants to interract
with the streams explicitly request that.