CLI: Add `--limit-request`
Note that this flag is the *total request size*, not the *message body size*.
I wanted the default value to be a round number (`1MiB`) rather than `1048448B` (`1MiB - 128B`).

When `--development` is set, the default `--limit-request` is `32KiB`.
When `--development` is _not_ set, the default is the full message size (`1MiB`).

`--development`'s choice is somewhat arbitrary, but my considerations were:

1. It brings a single-replica cluster with `--development` down below `1GiB` RSS.
2. Subsequent decreases have diminishing returns reducing RSS:
  - `--development --limit-request=64KiB` allocates `1029MiB`
  - `--development --limit-request=32KiB` allocates `994MiB`
  - `--development --limit-request=16KiB` allocates `977MiB`
  - `--development --limit-request=8KiB` allocates `968MiB`