CLI: Reduce `--limit-pipeline-requests` for production replica
Our replicas have a `clients_max` of 64, as a margin of safety.
(It impacts storage, so we overestimate the requirement).

But in practice, if users are batching properly, they probably have far fewer than 64 clients.
So we could reduce the default value of `--limit-pipeline-requests`, (and allow users who need more to set the flag manually).
This reduces production replica RSS. (For example, if we set `--limit-pipeline-requests=24` (for a total pipeline size of 32) then RSS drops by ~190MiB.)

The choice of `24` (for a total pipeline size of `32`) is somewhat arbitrary.
That is probably still on the high side, but I think we still want our default to overestimate what the typical user will need, to leave a margin.