state_machine: allow useless pulses
Primary decides whether to pulse or not, based on it's know (or unknown)
next transfer to expire. But in general different replicas might have
different state knowledge about pending expiring transfer.

So, a primary might pulse because it _doesn't_ know, but then the backup
would know _not_ to pulse, and yet it sill must process the pulse.

So we definietly should remove the assert, and instead:

* either don't do the scan at all
* or assert that it returned zero results

This PR take the second option as it makes assertions denser.

Seed:  ./zig/zig build  simulator_run -- --lite 11775367263850621025