cfo: successful seeds are displayed for pull request
When fuzzing my PRs, I've found myself double-checking if I used correct
branch names. To make the set of fuzzers run for a PR more obvious,
let's include successful seeds from PRs and color-code them green.

I debated whether this should be done for the main branch as well, but I
think it's better not to, as the number of fuzzers in general can get
rather large!

I also debated whether I should stop showing canary entries, but I think
it's better to keep those, as the purpose is different:

* canary demonstrates that the fuzzer is capable of failing
* the green seeds demonstrate the narrow set of fuzzers used for a PR