state_machine: remove pulse reset
There's no need to reset the pulse on the primary. What we are resetting
here is "the earliest next transfer to expire", but backups have that
up-to-date!

Where we do need a reset is after state-sync or restoring from
checkpoint, and we have that by virtue of the entire state machine
reset.