vopr: trivial bug in workload
transfers_delivered_recently is a heap. Its `items` refers to the entire
allocation, an not just to the initialized part. That is, there's
    transfers_delivered_recently.len
and
    transfers_delivered_recently.items.len

So we are essentially poking private implementation details here, and
getting what we've asked for (zero-initialized data).

Sometimes one can wish for private fields in Zig indeed!

Seed: ./zig/zig build simulator_run -- --lite 3712666001180766802
Closes: #1944