vopr: tics_max are runtime-overridable
Wanted to do this since forever, as bisecting on the ticks_max is the
first step to debug a liveness issue. Now that we use flags in the
simulator, we can just go and do it.

This is a small step -- a more principled approach would be perhaps to
stuff the entire config in the CLI, so that everything is overidable.
However, I rarely need to override anything but the ticks, so lets start
small!