flags: add tests for underscores in numbers
For numerical CLI arguments, we support readability underscores like
10_000_000. We get for free from Zig's std.parseInt function.

At the time of writing, this wasn't intentional, but we might as well
officially support this (I was actually thinking about _adding_ this
feature the other day, this is very useful for ticks_max argument of the
simulator).

So let's add tests here to make sure this won't regress in the future!

Note that underscores are _not_ supported for byte sizes. It won't be
hard to add support for `_`` for sizes, but we already support scale
suffixes, so adding underscores is in some sence redundant