benchmark: the devhub file size line was not flat
It is now flat.

...

Previously, the benchmark would calculate a target TPS and busy-wait to try
and achieve it.

The problem with that approach is it leads to unpredictable and random batch
sizing. Instead, remove the `transfer_count_per_second‎` flag entirely and
replace it with two new options:

* `transfer_batch_size` - sets the size of a transfer batch,
* `transfer_batch_delay_us` - sets the inter-batch delay.

These two flags allow obtaining the same behaviour, in a more predictable
and application style way, rather than trying to target an arbitrary TPS.
They default to 'as fast as possible'.

The random batch sizing lead to the datafile size varying with each run.

Also, fix that `account_balances` CLI flag wasn't being passed through
because we were setting defaults twice!