LSM/SetAssociativeCache: More granular value_count_max_multiple
Compute `SetAssociativeCache.value_count_max_multiple` more granularly.

`SetAssociativeCache` is used for:
- The grid cache.
- The object cache of each groove.

So `value_count_max_multiple` has two jobs:
- For each cache, the capacity (value count) must be a multiple of `value_count_max_multiple`.
- For the grid cache, the `value_count_max_multiple` is the _minimum_ cache capacity. (Whereas groove object caches are permitted to be zero.)

(More specifically, the purpose of `value_count_max_multiple` is to ensure that `SetAssociativeCache` CPU cache usage is optimal by ensuring that clock values (just a few bits) don't span multiple cache lines. (We already cheat at this to shrink our grid cache lower bound -- the grid pretends that it has a cache line size of 16.))

In order to shrink the replica's lower-bound memory usage, this value needs to shrink for the grid cache.

This (more complicated) calculation has the same results as the old one, for each cache.
But since the components are split apart, it is easier to modify (in the next commit).