multiversioning/flags: gate most flags behind --experimental
This makes it required to specify `--experimental` when starting TigerBeetle
if you are using flags that aren't explicitly documented in the help message.

Stable flags for `start` are specified in an allowlist - everything else is
considered experimental otherwise.

The main reason behind this is multiversion upgrades: by starting with a small
and intentionally limited set of flags, we can limit the surface area of what
we need to keep stable to ensure different versions keep compatibility with
the same CLI.