devhub: hide merged pull requests
For the main branch, we don't expect to see many fuzz failures and can
just wait for them to get evicted naturally.

For pull requests, we expect quite a bit more failures, so it makes
sense to proactively evict already merged PRs.

This is easy to do, because GitHub allows querying the list of PRs
without auth!