cfo: avoid useless compilations
Before this commit, CFO used to proactively compile all fuzzers for all
branches. After this change, only the required fuzzers are compiled.

Originally, the code was written with a clean split between "here we
build all the stuff we need" and "here we run all the stuff we need".
But, as the "stuff" became more dynamic (specifically, when we start
fuzzing labled PRs), the building logic got less precise: it was still
building everything for every branch, even if only a single fuzzer is
needed.

So instead run a separate build command for a specific fuzzer just
before we run it. While we could just run the fuzzer directly and rely
on zig to do the build automatically, this'll give us less precise
timings.

While we are at it, add a somewhat more clear logic about the fuzzing
plan for a parictular run of the CFO.