lsm: unit-test SortedSegmentedArray search bias
In the case of duplicates, many operations of SSA are left-biased. This
is somewhat non-obvious, as we don't explicitly ask for bias in the API,
and it feels like this might subtly regress in the future.

Let's add an explicit unit-test here, which nails down the current
behavior. Not saying that we shouldn't change it (the current API does
look suspicions to me), but let's make sure that if any change happens
here, it is intentional.