lsm: fix fuzz false positives
The check fails if there are several intMaxes in the array. I believe
that the check bellow for 0 is already correct in the face of the
duplicates.

See also the previous commit where I added explicit unit-tests for the
behavior in the face of duplicates.

Seed: ./zig/zig build  fuzz -- lsm_segmented_array 5619264251929892618