vsr: unify checking for production config
"Is this a production config?" is a somewhat ill-formed question, as,
physically, a config is just a set of many _specific_ options rather
than a broad toggle. But there are a couple of places where we need an
answer to that question to enable better sanity checking! This PR
introduces a specific function for this check, but it _should not_ be
used in "real" code --- instead, check for specific properties like
`config.verify`.

The main motivation here is preventing benchmark from poking into
`vsr_options`, which would allow us to simplify build.zig