build: use LazyPath for tb_client.h
Leverages LazyPath to propagate build dependencies for generating
tb_client.h.  The advantage of using LazyPath is it automatically
manages build step dependencies that use the lazy path.

In additon, this change adds a new InstallFile step based on the one
from std.  InstallFile from std is only meant to install files to the
install prefix, however, we want to install files to the repository
itself, which our new step can do.  We use this to "install" the
generated tb_client.h file to various locations.