build: add -Denforce-no-changes option
This option modifies the behavior of the build to enforce that any
generated files committed to the repo are not changed.  This change also
checks if the CI environment variable is set to "true" (which should be
the case on GitHub actions) and will enable this option by default when
it is.