test: merge config.test_min and config.fuzz_min
This PR reduces comptime-dimensionality of TigerBeetle by unifying
test_min and fuzz_min.

The two configs differ in the size of the storage: test_min is 200MiBs,
while fuzz_min is 1GiB, to allow hitting the lowest level of LSM tree.
Historically, this difference required a comptime split, because the
total size of the data file was a comptime parameter.

Nowadays though the storage size is elastic, and storage_size_limit is a
runtime parameter, so we no longer need to different comptime configs.

Implementation notes:

* bump test_min storage_limit to 1GiB, so that it can accomodate fuzzers.
* Don't change fuzzers: they use `constants.storage_size_limit_max` to
  pick the runtime size of the storage, which stays the same.
* Change the VOPR and replica test to explicitly request smaller grid
  sizes.
* FaultAtlas allocates a static bitsef of faults, which grows to 1GiB /
  sector_size, which is 2MiB, which feels acceptable.