vsr: don't crash in op_head_certain after state sync
The seed here crashes at

// Immediately after recovery, any faulty non-reserved prepares must be within our
// current checkpoint. See recovery case @M.

What happens here is that the statement is true after we recover the
journal from disk, but not after we used our vsr headers to further path
the journal.

Specifically, journal marks op=2 (which is outside of the checkpoint) as
case @E, and reserves it. But then we go and repair a vsr_headers header
over this slot. So it is:

* faulty
* dirty
* non-reserved
* in the past checkpoint

---

Moreover, just _after_ recovering the journal there's a stronger
statement to be made: all faulty headers are reserved. So let's assert
that instead

Seed: ./zig/zig build  vopr -- --lite 16548263080520932250