ci: sanity-check level coverage
It's easy to just run kcov, so let's just do that without thinking to
much about how it's best to integrate this into our overall dev process.

For now, let's just run a handful of test on main post merge, and upload
the results as a GitHub artifact (so, you can download the HTML report
manually).