go: remove concurrencyMax & fix pinning
Now that tb_client no longer allocates packets, it's Go's job.
Along with this, fix the CGO pinning error by having Go allocate
packet and not have a self-referential loop.

Unfortunately couldn't figure out how to stuff the packet
inside request to save on it's (escaped) allocation, but this works
for now.