build: unbreak CPU feature selection
This fixes the issue when, after Zig 0.13, TigerBeetle would sometimes
fall back to software implementaiton of AEGIS.

What we need to do is to just enable a couple of CPU features.
Previously, the logic worked by first performing normal target
resolution in build.zig, and then patching the target to what we need.

However, after 0.13 the patching no longer works: the features is
enabled in build.zig, but not in the actual binary. This is _probably_
due to discrepancy between Target.Query and ResolvedTarget, but I wasn't
able to trace this to the root.

Instead, the entire logic is re-written to just pick one of the set of
hard-coded targets.