VSR/Replica: Fix client deadlock due to register-evict-register
Scenario:

1. Client `A` sends an `operation=register` to a fresh cluster. (`A₁`)
2. Cluster prepares + commits `A₁`, and sends the reply to `A`.
4. `A` receives the reply to `A₁`, and issues a second request (`A₂`).
5. `clients_max` other clients register, evicting `A`'s session.
6. An old retry (or replay) of `A₁` arrives at the cluster.
7. `A₁` is committed (for a second time, as a different op, evicting one of the other clients).
8. `A` sends a second request (`A₂`), but `A` has the session number from the first time `A₁` was committed.

(To be clear, enrolling so many clients _is_ an operator error.
But we should handle it more gracefully nonetheless.)

Formerly in this case, the client `A` would be stuck forever (or at least until it is evicted for unrelated reasons, due to many other new clients registering).

Instead, send an eviction, so that the error can be bubbled up to the operator.