client: avoid synchronizing via user-exposed object
In languages like C# and Java, we can use objects additionally
as mutexes. Here, we use it to synchronize access to the tb_client
handle.

The use of the NativeClient object as the mutex means a library
user could accidently lock the NativeClient and interfere with
tb_client synchronization. Instead, NativeClient now holds its
own internal object to acquire locks on.