vsr: don't block commit on repairing ops  from the past checkpoint
Before this commit, we waiting until all repairable headers are fixed
before trying to commits prepares. This is suboptimal: if we have all
the header after the checkpoint, then we can start committing even if
some headers before checkpoint are missing. This PR does just that.

That seems like a good idea in general, to maximize availability, and to
not waste durability, but this also fixes a VOPR failure:

* Cluster of A, B, C, with C partitioned
* A and B move ahead
* B gets partitioned
* C joins A and state syncs to its checkpoint
* A gets corruption in the log in the previous checkpoint

This gets A into an interesting position --- in theory, it could commit
everything, because it has the checkpoing and everything after that. But
it gets stuck in repair, trying to fix items from the last checkpoint!
Formally speaking, the cluster is allowed to be unavailable here: with C
partitioned and A corrupted, we have two faults.  However, currently
this doesn't trigger "core missing prepare" case in the VOPR, because
the missing stuff is before A's commit_min.

No seed, as that was found on one of my temporary branches with
semi-random changes on top.