Intercept Ctrl-C during interactive REPL prompt
This allows users to cancel their current prompt without attempting to
parse it as a command. EOF is still needed to exit the REPL cleanly
while at the prompt.

Ctrl-C can still be used to kill the application when not in prompt mode
in the event of a hang/slow request.