vsr: new state sync protocol
This commit implements the new state sync protocol, which fixes a couple
of long-standing liveness issues. As an unintended bonus, the new
implementation ends up being simpler, at least if measured in lines of
code!

 ## Old Protocol Problems

The original thinking was to implement state-sync as a separate
protocol, mostly orthogonal to the rest of VSR. VSR maintains the log,
state sync maintains the checkpoint. Checkpoints are committed, so
there's no doubt regarding _what_ to sync, so a separate protocol sounds
easy enough, right?

Well, turns out its' not that simple, as the two issues were discovered:

_First_, the log and the checkpoint might become physically disjoint
(checkpoint ahead of the log). The implementation fixed this case by
maneuvering the replica into recovering_head, but that sadly created a
liveness issue --- recovering_head replicas don't participate in
consensus, so it is possible for cluster to get stuck (with most of the
cluster in recovering_head) even without any storage faults!

_Second_, the log and the checkpoint might become logically inconsistent
(the message in the log for the checkpoint's slot might be from the
different view). Although checkpoint is committed, that's not true about
the log! This issue manifested itself as assertion failures in
`valid_hash_chain`. There were several band-aid fixes for the failures,
which boiled down to maneuvering replica into recovering_head if
checkpoint's view and log view are different. But we weren't able to
plug all the wholes, as it slowly became apparent that fixing everything
would require duplicating VSR-specfic SV handling logic when dealing
with checkpoints.

 ## The New Protocol

The idea is rather simple --- add primary's checkpoint to SV and handle
state sync in `on_start_view`. This makes it so that the checkpoint and
log are consistent by construction. This also yields an unexpected
benefit: a replica no longer ignores SV's which are too far in the
future, it is always possible to install a header from SV (although you
might need to update the checkpoint as well).

The implementation is somewhat tricky --- we update checkpoint
asynchronously, but we update replica state immediately, so there's a
window while our `op` is consistent with in-memory checkpoint, but not
on-disc checkpoint. This signalled by

    self.syncing == .updating_superblock

state. While in this state, the replica mostly just waits for the
superblock update to finish and doesn't do much else.

Additionally, the implementation right now is optimized for smaller
diff, rather than for path dependent perfection. For example, we still
call it `Superblock.view_change`, even though now this function can be
called _only_ to advance the checkpoint state. I want to do a follow up
cleanup PR. There's perhaps more code/asserts that needs to be adjusted
now that `.recovering_head` is again possible only after `.recovering`.

The equivalent code have been running in VOPR for quite some time, and I
think any sync-consistency related issues are fixed. The one current
outstading VOPR bug (with several seeds) is an orthogonal issue where we
allow to sync to a checkpoint before it's sync_content is done.

This particular commit has some extra assertions and cleanups, so they
might still blow up, but hopefully would be trivial to fix.