tidy: don't check client_readmes.zig for function length
We have some giant functions in that file, but they are essentially
templated markdown documents, so they are expected to be large.

I've tried extracting each individual section to a separate function,
and, while it is not too ugly and gets past the tidy, I wouldn't
consider that to be an actual improvement!

So let's opt out here!