vsr: prepare to bump protocol version for the new sync protocol
The recently introduced new sync protocol changes the semantics of SV
message, and necessitates incrementing vsr.Version. Naturally, we want
to provide a smooth upgrade path for existing clusters.

While we do have an upgrade path implemented for state machine changes,
we don't  have one for protocol changes.

This commit doesn't aim to provide a general solution just yet, and
instead focuses on the specific state sync change, which can be rolled
out relatively easily.

The protocol with version=0 is the protocol we use today. The protocol
with version=1 is the protocol which is identical to version=0 for all
messages except for SV.

In the next transitional release, replicas will  accept either version=0
or version=1 messages. They will send only version=0 messages, except
for an SV, which will be send in both versions (so, two different
messages). The release after that would be a normal releases, speaking
exclusively protocol version=1.