StateMachine: Fix balancing transfers with large balances
### Context

When a balancing transfer is created with `amount = 0`, the transfer should move as much money as possible (within the constraints of the respective accounts).

### Bug

Balancing transfers used on accounts with net balances greater than `2^64 - 1` would not move as much money as they should.

(There was a leftover `u64` from before amounts/balances were changed to `u128`.)