build: refactor client bindings
Our binding generated code survived a number of code motions, and
accumulated a bunch of accidental complexity. This new implementation
aims at simplifying the things overall.

Bindings are source files in target language (.java, .go, etc), which
are generated from our tigerbeetle.zig code. We want the generated files
to change automatically when the sources change. At the same time, we
want to commit the generated files to our repository. We expect this
files to be human readable. Users our our libraries should read
generated bindings in their language, rather then the generating code in
Zig.

This is achieved by generating the bindings in build.zig, with a twist:
if a CI env variable is set, instead of overwriting the files, they are
checked for freshness.

If the generator generates a single file, it should write it to stdout.

If it generates multiple files, it should generate them to the output
directory, which is specified as the sole CLI argument.

Implementation-wise, `Generated` struct handles all the details.