build: remove CopyFile
CopyFile plays a dual role:

* like `Generated`, it has the logic to update files in develompent, and
  ensure freshness on CI.
* like builtin `InstallStep`, it copies files around (unlike install, it
  does that outside of the perfix)

As we now have a dedicated `Generated`, and we _could_ abuse
`InstallStep` to install files where we want to, it feels like
`CopyFile` no longer pulls its weight. Although passing `../../` to
install to escape `zig-out/bin` is hacky, it works well enough, and we
are using it elsewhere anyway, so, for simplicity, lets stick with
standard install step.