multiversioning: don't timeout_start() twice
If `statx` returns an error, it would set some state, call `timeout_start()`
and then call `handle_error()`. Except, `handle_error()` itself calls
`timeout_start()` which leads to an assertion failure.

Practically, this means that upgrading 0.15.4 by deleting the binary and
replacing it was broken; overwriting in place for upgrades still works
(or anything that wouldn't cause the stat to fail).